refactor: Value Object 별도 폴더로 이관
